normalize weights as frequences

+++ extract_cfg.hpp
@@ -5,4 +5,4 @@
 #include <fstream>
 #include <string>
 
-std::map<std::string, std::vector<std::tuple<int, std::string> > > make_cfg(char* ir_file_name, char* profiling_file_name);
+std::map<std::string, std::vector<std::tuple<float, std::string> > > make_cfg(char* ir_file_name, char* profiling_file_name);
